refactor: Refatorar relacionamentos e classes

+++ desafio-tecnico/src/main/java/com/transferencia/desafio_tecnico/model/dtos/transferencia/TransferenciaResquestDto.java
@@ -19,7 +19,7 @@ public class TransferenciaResquestDto {
 
     private Long recebedor;
 
-    private Double valor;
+    private BigDecimal valor;
 
     private TipoTransacao tipoTransacao;
